requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aioredis==1.1.0 ; python_version > "3.3"
arrow==0.12.1
async-timeout==2.0.0 ; python_version > "3.3"
attrs==18.1.0
backports.functools-lru-cache==1.5  # via arrow
enum34==1.1.6 ; python_version < "3.4"
hiredis==0.2.0            # via aioredis
python-dateutil==2.7.3    # via arrow
redis==2.10.6
simplejson==3.16.0
six==1.11.0               # via python-dateutil, structlog
sqlalchemy==1.2.10
structlog==18.1.0
typing==3.6.4 ; python_version < "3"
